bot/requirements.txt
aiohttp
asyncio
discord.py
sqlalchemy